+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- 
+- Update vue-components library version and fix to new version 1.0.15
+- Implemented 2 events that fires on page render `page_payload_preprocess` and `page_payload_postprocess` and that can be listened to by any external js code: Foundation for Js events firing + fire on page render #166
 
 
 ### [1.1.1] - 2025-03-28